pyproject.toml
[tool.poetry]
name = "pythoneda-artifact-changes"
version = "0.0.1a1"
description = "Artifact space in charge of managing source code changes"
authors = ["rydnr <github@acm-sl.org>"]
readme = "README.md"
homepage = "https://github.com/pythoneda-artifact/changes"
packages = [{include = "pythonedaartifactchanges"}]

[tool.poetry.dependencies]
dulwich = "^0.21.5"
GitPython = "^3.1.27"
paramiko = "^2.11.0"
python = "^3.8"
pythoneda-base = "^0.0.1a15"
pythoneda-shared-git = "^0.0.1a3"
pythoneda-artifact-event-changes = "^0.0.1a1"
semver = "^3.0.0"

[tool.poetry.dev-dependencies]
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
